Additions, replacements, and fixes (oh my!)
Several files for jobs I don't play or understand (yet, in the case of
RPR) are just those of the original I forked, modified just enough to
compile. For the others, I manually integrated some changes, tweaked the
RDM stuff, and decided to just throw it all into a single update.