MCH HS/AA/D feature improvements
The HotShot/AirAnchor/Drill feature for machinist should now try to
prioritise between Drill and Air Anchor (but only between those two)
based on your battery gauge, to prevent overcapping. Drill is still
preferred over Hot Shot for the potency increase. If you enable the
Chainsaw addition, Air Anchor will be prioritised when all three are
available AND your battery gauge is no more than 80. Otherwise, it'll
choose between Chainsaw and Drill for priority based on the original
action as before.

This is a /slightly/ experimental update, but some brief testing on my
own game seems to be working fine, so barring any unexpected issues with
level sync or something, it shouldn't be an issue.