Tweak MCH HS/AA/D+CS
The HS/AA/D+CS feature no longer changes Chain Saw, so that it can be kept for on-demand use. It will (if enabled) be added to the existing replacer as the last priority (only when Drill and Air Anchor are both on cooldown) since it's an AoE rather than ST, which may not always be desirable.