Bump version to release all the changes made
I'm so tired. You may not know, but this release involves TWENTY ONE
commits. I have MANUALLY gone over EVERY JOB'S COMBOS to compare my
version with daemitus's original fork that mine came from, in order
to find things that are on his that I needed to merge in. And there's
more to do soon, too. I've decided on some changes to the style of
the combo functions, I'm going to have to go over the whole project
AGAIN at some point.

The things I do for my projects.

It's fine, I enjoy it after the fact :P