Refactoring and update alerts
Several parts have been moved to better organise the workspace,
including separating out features that are primarily based on
internal game data, as well as the core of the configuration system
in preparation for the upcoming detail settings update.

Additionally, an automatic chat alert for plugin updates has been
implemented, and can be disabled in the settings window if desired.
The intent of this is to help people not miss new or significantly
changed features by letting them know that things have been changed,
and providing an easy way to open the settings window immediately.