New and fixed PLD Confiteor combos
The Paladin Confiteor feature was also turning most but not all actions
it applied to into the Blade of Faith/Truth/Valor chain after Confiteor.
This has been extracted out into a new child combo and should now apply
to all actions, closing #66.

On the other hand, it's late, I'm tired, and I'll be going to sleep in
twenty minutes, so I'm pretty sure it's all good but if I wake up to a
report about it being broken, I'll know I fucked up!