Minor bugfix for RDM smartcast ST melee combo
The ST smartcast spell button now becomes the melee combo if both mana
levels are at 100, in case you cap your gauge entirely. It's not an
"optimal" way to do the melee, but it allows an immediate second run
that is "optimal" and it prevents overcapping, not to mention the only
way to even it out without doing the melee combo would be to cast a
couple of enchanted reprises then another ver* cycle, and at that point
it's probably better to just double melee+finisher instead.