Tweak the RDM smartcast ST algorithm a little
Mostly new options for when to become Acceleration and Swiftcast, but
also a slight (hopefully an) improvement to the decision to enter the
melee combo. If you're already mid-combo, it'll force into the combo,
assuming you're in melee range; otherwise, it'll only enter the combo
if you're not accelerated or fastcasting so as to save those.