Completely rewrite RDM single target smartcast
The entire algorithm has changed under the hood. Everything has been
restructured for easier development and to improve the prioritisation
of certain actions (Acceleration/Swiftcast when weaving/moving will now
be done first, before all other options) in order to minimise the casting
of Jolt (II) as much as possible. This is technically slightly experimental
but twenty minutes of testing on a training dummy at my FC's estate has
revealed no problems so far, so I'm calling it releasable.