Add new SGE combos
New combo "Flying Phlegma" is the reverse of the "Phlemga Gap Closer"
combo, turning Icarus into Phlegma when it's up AND you're in range to
use it. No configuration this time.

There are three new combos for Dosis, turning it into Phlegma, Toxikon,
or Dyskrasia (in that order or priority, if all are enabled) when you're
moving and Eukrasia is NOT active. As usual, it won't become actions
that you can't use, so if you have no charges of Phlegma, that'll be
skipped, and so will Toxikon if you have no Addersting, etc.