Add DNC Starfall Dance saver feature
Independently available and configurable for single-target and AoE
multibuttons, if you have Flourishing Starfall and the time left is
no more than the configured threshold, the multibutton in question
becomes Starfall Dance to prevent losing the chance to use it.

Set to 20 seconds (maximum buff time) to use immediately, set to 3
seconds (minimum time to ensure you don't lose your chance to the GCD
cooldown) to use only at the last metaphorical second before it's lost.