New MCH features
Barrel Stabilizer weaving on the main combo, Hypercharge weaving on the same when heat gauge reaches a certain amount, smart overriding of the same when Reassembled, Gauss/Riccochet can now weave directly onto Heat Blast and the main combo HB override will incorporate it as well, and Tactician/Dismantle will swap out depending on level, cooldown, and whether you have a target or not.

I should've split this up, but I coded a storm all at once and I'm not pulling the diff apart for it.